rename misnamed archive button component
