+++ app/frontend/components/domains/jurisdictions/users/index.tsx
@@ -12,7 +12,7 @@ import { Paginator } from "../../../shared/base/inputs/paginator"
 import { PerPageSelect } from "../../../shared/base/inputs/per-page-select"
 import { LoadingScreen } from "../../../shared/base/loading-screen"
 import { SharedSpinner } from "../../../shared/base/shared-spinner"
-import { ToggleArchivedButton } from "../../../shared/buttons/show-archived-button"
+import { ToggleArchivedButton } from "../../../shared/buttons/toggle-archived-button"
 import { SearchGrid } from "../../../shared/grid/search-grid"
 import { SearchGridItem } from "../../../shared/grid/search-grid-item"
 import { RouterLinkButton } from "../../../shared/navigation/router-link-button"
