+++ app/frontend/components/domains/requirement-template/screens/requirement-template-screen.tsx
@@ -3,7 +3,7 @@ import { observer } from "mobx-react-lite"
 import React from "react"
 import { useTranslation } from "react-i18next"
 import { useMst } from "../../../../setup/root"
-import { ToggleArchivedButton } from "../../../shared/buttons/show-archived-button"
+import { ToggleArchivedButton } from "../../../shared/buttons/toggle-archived-button"
 import { RouterLinkButton } from "../../../shared/navigation/router-link-button"
 import { RequirementTemplateGrid } from "../../../shared/requirement-template/requirement-template-grid"
 import { TemplateVersionsSidebar } from "../template-versions-sidebar"
