+++ app/frontend/components/domains/super-admin/early-access/requirement-templates/index.tsx
@@ -13,7 +13,7 @@ import { CustomMessageBox } from "../../../../shared/base/custom-message-box"
 import { Paginator } from "../../../../shared/base/inputs/paginator"
 import { PerPageSelect } from "../../../../shared/base/inputs/per-page-select"
 import { SharedSpinner } from "../../../../shared/base/shared-spinner"
-import { ToggleArchivedButton } from "../../../../shared/buttons/show-archived-button"
+import { ToggleArchivedButton } from "../../../../shared/buttons/toggle-archived-button"
 import { SearchGrid } from "../../../../shared/grid/search-grid"
 import { SearchGridItem } from "../../../../shared/grid/search-grid-item"
 import { RouterLinkButton } from "../../../../shared/navigation/router-link-button"
