+++ app/frontend/components/domains/super-admin/site-configuration-management/users-screen.tsx
@@ -9,7 +9,7 @@ import { useMst } from "../../../../setup/root"
 import { Paginator } from "../../../shared/base/inputs/paginator"
 import { PerPageSelect } from "../../../shared/base/inputs/per-page-select"
 import { SharedSpinner } from "../../../shared/base/shared-spinner"
-import { ToggleArchivedButton } from "../../../shared/buttons/show-archived-button"
+import { ToggleArchivedButton } from "../../../shared/buttons/toggle-archived-button"
 import { SearchGrid } from "../../../shared/grid/search-grid"
 import { SearchGridItem } from "../../../shared/grid/search-grid-item"
 import { RouterLinkButton } from "../../../shared/navigation/router-link-button"
